+++ internal/cli/root.go
@@ -21,6 +21,7 @@ import (
 	"strings"
 	"tidbcloud-cli/internal/cli/branch"
 	"tidbcloud-cli/internal/cli/serverless"
+	"tidbcloud-cli/internal/cli/upgrade"
 
 	"tidbcloud-cli/internal"
 	configCmd "tidbcloud-cli/internal/cli/config"
@@ -28,7 +29,6 @@ import (
 	"tidbcloud-cli/internal/cli/dataimport"
 	"tidbcloud-cli/internal/cli/dataimport/start"
 	"tidbcloud-cli/internal/cli/project"
-	"tidbcloud-cli/internal/cli/update"
 	"tidbcloud-cli/internal/cli/version"
 	"tidbcloud-cli/internal/config"
 	"tidbcloud-cli/internal/flag"
@@ -168,7 +168,7 @@ func RootCmd(h *internal.Helper) *cobra.Command {
 	rootCmd.AddCommand(configCmd.ConfigCmd(h))
 	rootCmd.AddCommand(project.ProjectCmd(h))
 	rootCmd.AddCommand(version.VersionCmd(h))
-	rootCmd.AddCommand(update.UpdateCmd(h))
+	rootCmd.AddCommand(upgrade.Cmd(h))
 	rootCmd.AddCommand(dataimport.ImportCmd(h))
 	rootCmd.AddCommand(connect.ConnectCmd(h))
 	rootCmd.AddCommand(branch.BranchCmd(h))
