+++ internal/cli/upgrade/upgrade.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package update
+package upgrade
 
 import (
 	"bytes"
@@ -35,10 +35,10 @@ import (
 	"go.uber.org/zap"
 )
 
-func UpdateCmd(h *internal.Helper) *cobra.Command {
+func Cmd(h *internal.Helper) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "update",
-		Short: "Update the CLI to the latest version",
+		Use:   "upgrade",
+		Short: "upgrade the CLI to the latest version",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			ctx := cmd.Context()
 			// If is managed by TiUP, we should disable the update command since binpath is different.
